BlackJack / historico pessoa

+++ ProjetoPOO/Uteis.cpp
@@ -9,7 +9,7 @@ int Util::RandNumInt(int LimInf, int LimSup)
 	return dis(gen);
 }
 
-double Util::RandNumDouble(int LimInf, int LimSup)
+double Util::RandNumDouble(double LimInf, double LimSup)
 {
 	std::random_device rd;
 	std::mt19937 gen(rd());
